Move implementations of nalgebra trait into macros
