+++ src/datatypes/dual3.rs
@@ -122,3 +122,4 @@ impl<T: fmt::Display, F> fmt::Display for Dual3<T, F> {
 
 impl_third_derivatives!(Dual3, [v1, v2, v3]);
 impl_dual!(Dual3, [v1, v2, v3]);
+impl_nalgebra!(Dual3, [v1, v2, v3]);
